Move CLI helper functions to dedicated module
